requirements.txt
# requirements
configparser
docker-py==1.10.6